Feature/resident

+++ src/main/java/com/hust/smart_apartment/service/impl/ApartmentServiceImpl.java
@@ -1,13 +1,19 @@
 package com.hust.smart_apartment.service.impl;
 
+import com.hust.smart_apartment.constants.LivingType;
 import com.hust.smart_apartment.dto.ModifyDto;
 import com.hust.smart_apartment.dto.request.ApartmentRequest;
 import com.hust.smart_apartment.dto.request.SearchRequest;
+import com.hust.smart_apartment.dto.request.UpdateApartmentRequest;
 import com.hust.smart_apartment.dto.response.ApartmentResponse;
 import com.hust.smart_apartment.entity.Apartment;
+import com.hust.smart_apartment.entity.Floor;
+import com.hust.smart_apartment.entity.Resident;
+import com.hust.smart_apartment.entity.ResidentChangeLog;
 import com.hust.smart_apartment.mapper.ApartmentMapper;
 import com.hust.smart_apartment.mapper.ResidentMapper;
 import com.hust.smart_apartment.repository.ApartmentRepository;
+import com.hust.smart_apartment.repository.FloorRepository;
 import com.hust.smart_apartment.repository.ResidentRepository;
 import com.hust.smart_apartment.repository.SearchRepository;
 import com.hust.smart_apartment.service.ApartmentService;
@@ -15,6 +21,12 @@ import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -25,9 +37,10 @@ public class ApartmentServiceImpl implements ApartmentService {
     private final ApartmentMapper apartmentMapper;
     private final ResidentMapper residentMapper;
     private final SearchRepository<ApartmentResponse> searchRepository;
+    private final FloorRepository floorRepository;
+
     @Override
     public ApartmentResponse getById(Long id) {
-        // Fetch apartment by ID or throw an exception if not found
         Apartment apartment = apartmentRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Apartment not found with ID: " + id));
         return apartmentMapper.entityToResponse(apartment);
@@ -35,22 +48,48 @@ public class ApartmentServiceImpl implements ApartmentService {
 
     @Override
     public ApartmentResponse create(ApartmentRequest request) {
-        // Map request DTO to entity, save to DB, and map the result back to response DTO
+
         Apartment apartment = apartmentMapper.requestToEntity(request);
-        Apartment savedApartment = apartmentRepository.save(apartment);
-        return apartmentMapper.entityToResponse(savedApartment);
+        Floor floor = floorRepository.findById(request.getFloorId())
+                .orElseThrow(() -> new EntityNotFoundException("Floor not found with ID: " + request.getFloorId()));
+        apartment.setFloor(floor);
+        return apartmentMapper.entityToResponse(apartmentRepository.save(apartment));
     }
 
+
     @Override
-    public ApartmentResponse update(Long id, ApartmentRequest request) {
-        // Fetch apartment by ID, update fields, and save the updated entity
+    public ApartmentResponse update(Long id, UpdateApartmentRequest request) {
         Apartment existingApartment = apartmentRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Apartment not found with ID: " + id));
 
-        // Update fields (assuming mapper can do this with a helper method)
-        existingApartment.setCode(request.getCode());
-        existingApartment.setName(request.getName());
-        existingApartment.setArea(request.getArea());
+        Resident owner = request.getOwner();
+        List<Resident> residents = request.getResidents().stream().map(residentMapper::requestToEntity).toList();
+        Set<Long> oldResidentIds = existingApartment.getResidents().stream().map(Resident::getResidentId).collect(Collectors.toSet());
+        Set<Resident> createdResidentIds = residents.stream().filter(resident -> !oldResidentIds.contains(resident.getResidentId())).collect(Collectors.toSet());
+        Set<Resident> updateResidentIds = residents.stream().filter(resident -> oldResidentIds.contains(resident.getResidentId())).collect(Collectors.toSet());
+        List<Resident> removeResidentIds = existingApartment.getResidents().stream().filter(resident -> !updateResidentIds.contains(resident)).peek(x -> x.setLivingApartment(null)).toList();
+
+        List<Resident> residentList = new ArrayList<>(residentRepository.saveAll(updateResidentIds));
+        residentList.addAll(residentRepository.saveAll(createdResidentIds));
+        residentList.addAll(residentRepository.saveAll(removeResidentIds));
+
+//        List<ResidentChangeLog> logs = residentList.stream().map(resident -> {
+//            ResidentChangeLog log = new ResidentChangeLog();
+//            log.setApartment(apartmentMapper.entityToResponse(existingApartment));
+//            log.setResident(residentMapper.entityToResponse(resident));
+//            log.setChangeDate(LocalDateTime.now());
+//            log.setNotes("Change owner");
+//            if(createdResidentIds.contains(resident)) {
+//                log.setLastType(LivingType.O_NGOAI);
+//                log.setChangeType(resident.getCurrentLivingType());
+//            }else {
+//                log.setLastType(resident.getCurrentLivingType());
+//                log.setChangeType(LivingType.O_NGOAI);
+//            }
+//        });
+
+        existingApartment.setOwner(owner);
+        existingApartment.setResidents(residents);
         Apartment updatedApartment = apartmentRepository.save(existingApartment);
 
         return apartmentMapper.entityToResponse(updatedApartment);
@@ -58,10 +97,9 @@ public class ApartmentServiceImpl implements ApartmentService {
 
     @Override
     public ModifyDto delete(Long id) {
-        // Fetch apartment by ID, delete it, and return success status
-        apartmentRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Apartment not found with ID: " + id));
-
+        if (!apartmentRepository.existsById(id)) {
+            throw new EntityNotFoundException("Apartment not found with ID: " + id);
+        }
         return new ModifyDto(apartmentRepository.deleteApartmentByApartmentId(id));
     }
